requirements.txt
numpy>=1.14.0
gdown>=3.10.1
Pillow>=5.2.0
opencv-python>=4.5.5.64
tensorflow>=1.9.0
keras>=2.2.0
mtcnn>=0.1.0
retina-face>=0.0.1
opencv-contrib-python>=4.3.0.36
mediapipe>=0.8.7.3
dlib>=19.20.0
patool>=1.2
ultralytics>=8.0.122